# docs/api/swagger.yml
openapi: 3.0.0
info:
  version: "1.0.0-SNAPSHOT-oas3"
  title: OverHours REST API
  description: ''
...
paths:
...
  '/meetings/{meetingId}':
    get:
      tags:
        - meetings
      summary: Get meeting with ID
      parameters:
        - name: meetingId
          in: path
          description: ID of meeting to return
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Meeting'
        '404':
          description: not found
    put:
      tags:
        - meetings
      summary: Update a meeting with ID
      parameters:
        - name: meetingId
          in: path
          description: ID of meeting to update
          required: true
          schema:
            type: string
      responses:
        '201':
          description: Successfully update a Meeting
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChangeInfo'
        '400':
          description: Invalid input
      requestBody:
        description: A JSON object of TimeLog
        content:
          application/json:
            schema:
              type: object
              properties:
                StartTime:
                  type: integer
                SeasonId:
                  type: string
    delete:
      tags:
        - meetings
      summary: Delete a meeting with ID
      parameters:
        - name: meetingId
          in: path
          description: ID of meeting to delete
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Successfully delete a meeting
        '400':
          description: Invalid meetingId
        '404':
          description: not found
components:
  schemas:
...
    Meeting:
      type: object
      properties:
        MeetId:
          type: string
        StartTime:
          type: integer
        SeasonId:
          type: string
        Title:
          type: string
        Description:
          type: string
        CheckinLevel:
          type: integer
        StartCheckinTime:
          type: integer
        FinishTime:
          type: integer
        Participants:
          type: array
          items:
            $ref: '#/components/schemas/ParticipantData'
    ParticipantData:
      type: object
      properties:
        UserId:
          type: string
        Leave:
          type: boolean
        IsAdmin:
          type: boolean
    ChangeInfo:
      type: object
      properties:
        Updated:
          type: integer
        Removed:
          type: integer
        Matched:
          type: integer
        UpsertedId:
          type: string
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session_jwt
security:
  - cookieAuth: []
servers:
  - url: 'https://virtserver.swaggerhub.com/kennhung/OverHours/1.0.0-SNAPSHOT'
  - url: 'http://virtserver.swaggerhub.com/kennhung/OverHours/1.0.0-SNAPSHOT'